requirements.txt
flask==2.3.2
flask-mysql==1.5.2